Minor updates. Changed type Tickf to float instead of double to make it cross platform between 32 and 64 bit machines.

+++ trunk/Planet/src/basic.h
@@ -29,7 +29,7 @@ namespace Planet
 	typedef Vec3f Vec;
 	typedef Vec Pos;
 
-	typedef double Tickf;
+	typedef float Tickf;
 
 	//template <typename T>
 	inline float radianToDegree(float angle)
